+++ lib/async.dart
@@ -5,6 +5,7 @@
 library dart.pkg.async;
 
 export "result.dart";
+export "src/async_memoizer.dart";
 export "src/delegate/event_sink.dart";
 export "src/delegate/future.dart";
 export "src/delegate/sink.dart";