+++ CHANGELOG.md
@@ -10,6 +10,9 @@
 - Added `SubscriptionStream` which creates a single-subscription stream
   from an existing stream subscription.
 
+- Added a `ResultFuture` class for synchronously accessing the result of a
+  wrapped future.
+
 - Added `FutureGroup.onIdle` and `FutureGroup.isIdle`, which provide visibility
   into whether a group is actively waiting on any futures.